Feature/move logging (#1493)
* move logging provider out

* move logging provider to own directory, remove singleton

* cleanup
