+++ py/core/__init__.py
@@ -110,9 +110,7 @@ __all__ = [
     "LogFilterCriteria",
     "LogProcessor",
     # Logging Providers
-    "SqlitePersistentLoggingProvider",
-    "LoggingConfig",
-    "R2RLoggingProvider",
+    "PersistentLoggingConfig",
     # Run Manager
     "RunManager",
     "manage_run",
@@ -125,7 +123,6 @@ __all__ = [
     ## PIPES
     "AsyncPipe",
     "AsyncState",
-    "PipeType",
     ## PROVIDERS
     # Base provider classes
     "AppConfig",
