+++ py/core/base/__init__.py
@@ -82,10 +82,7 @@ __all__ = [
     "LogAnalyticsConfig",
     "LogFilterCriteria",
     "LogProcessor",
-    # Logging Providers
-    "SqlitePersistentLoggingProvider",
-    "LoggingConfig",
-    "R2RLoggingProvider",
+    "PersistentLoggingConfig",
     # Run Manager
     "RunManager",
     "manage_run",
@@ -98,7 +95,6 @@ __all__ = [
     ## PIPES
     "AsyncPipe",
     "AsyncState",
-    "PipeType",
     ## PROVIDERS
     # Base provider classes
     "AppConfig",
