+++ py/core/base/logging/__init__.py
@@ -6,11 +6,7 @@ from .log_processor import (
     LogFilterCriteria,
     LogProcessor,
 )
-from .r2r_logger import (
-    LoggingConfig,
-    R2RLoggingProvider,
-    SqlitePersistentLoggingProvider,
-)
+from .r2r_logger import PersistentLoggingConfig
 from .run_manager import RunManager, manage_run
 
 __all__ = [
@@ -22,9 +18,7 @@ __all__ = [
     "LogFilterCriteria",
     "LogProcessor",
     # Logging Providers
-    "SqlitePersistentLoggingProvider",
-    "LoggingConfig",
-    "R2RLoggingProvider",
+    "PersistentLoggingConfig",
     # Run Manager
     "RunManager",
     "manage_run",
