+++ py/core/base/pipes/__init__.py
@@ -1,3 +1,3 @@
-from .base_pipe import AsyncPipe, AsyncState, PipeType
+from .base_pipe import AsyncPipe, AsyncState
 
-__all__ = ["AsyncPipe", "AsyncState", "PipeType"]
+__all__ = ["AsyncPipe", "AsyncState"]
