+++ py/core/main/assembly/factory.py
@@ -14,10 +14,10 @@ from core.base import (
     EmbeddingProvider,
     IngestionConfig,
     OrchestrationConfig,
-    R2RLoggingProvider,
 )
 from core.pipelines import RAGPipeline, SearchPipeline
 from core.pipes import GeneratorPipe, MultiSearchPipe, SearchPipe
+from core.providers.logging.r2r_logging import SqlitePersistentLoggingProvider
 
 from ..abstractions import R2RAgents, R2RPipelines, R2RPipes, R2RProviders
 from ..config import R2RConfig
@@ -225,6 +225,9 @@ class R2RProviderFactory:
             Union[OpenAICompletionProvider, LiteLLMCompletionProvider]
         ] = None,
         orchestration_provider_override: Optional[Any] = None,
+        r2r_logging_provider_override: Optional[
+            SqlitePersistentLoggingProvider
+        ] = None,
         *args,
         **kwargs,
     ) -> R2RProviders:
@@ -272,6 +275,12 @@ class R2RProviderFactory:
             or self.create_orchestration_provider(self.config.orchestration)
         )
 
+        logging_provider = (
+            r2r_logging_provider_override
+            or SqlitePersistentLoggingProvider(self.config.logging)
+        )
+        await logging_provider.initialize()
+
         return R2RProviders(
             auth=auth_provider,
             database=database_provider,
@@ -279,6 +288,7 @@ class R2RProviderFactory:
             ingestion=ingestion_provider,
             llm=llm_provider,
             orchestration=orchestration_provider,
+            logging=logging_provider,
         )
 
 
@@ -350,6 +360,7 @@ class R2RPipeFactory:
         from core.pipes import ParsingPipe
 
         return ParsingPipe(
+            logging_provider=self.providers.logging,
             ingestion_provider=self.providers.ingestion,
             database_provider=self.providers.database,
             config=AsyncPipe.PipeConfig(name="parsing_pipe"),
@@ -362,6 +373,7 @@ class R2RPipeFactory:
         from core.pipes import EmbeddingPipe
 
         return EmbeddingPipe(
+            logging_provider=self.providers.logging,
             embedding_provider=self.providers.embedding,
             database_provider=self.providers.database,
             embedding_batch_size=self.config.embedding.batch_size,
@@ -375,6 +387,7 @@ class R2RPipeFactory:
         from core.pipes import VectorStoragePipe
 
         return VectorStoragePipe(
+            logging_provider=self.providers.logging,
             database_provider=self.providers.database,
             config=AsyncPipe.PipeConfig(name="vector_storage_pipe"),
         )
@@ -386,6 +399,7 @@ class R2RPipeFactory:
         from core.pipes import VectorSearchPipe
 
         return VectorSearchPipe(
+            logging_provider=self.providers.logging,
             database_provider=self.providers.database,
             embedding_provider=self.providers.embedding,
             config=SearchPipe.SearchConfig(name="vector_search_pipe"),
@@ -407,6 +421,7 @@ class R2RPipeFactory:
         )
 
         query_transform_pipe = QueryTransformPipe(
+            logging_provider=self.providers.logging,
             llm_provider=self.providers.llm,
             database_provider=self.providers.database,
             config=QueryTransformPipe.QueryTransformConfig(
@@ -416,6 +431,7 @@ class R2RPipeFactory:
         )
 
         return MultiSearchPipe(
+            logging_provider=self.providers.logging,
             query_transform_pipe=query_transform_pipe,
             inner_search_pipe=inner_search_pipe,
             config=multi_search_config,
@@ -446,6 +462,7 @@ class R2RPipeFactory:
         from core.pipes import RoutingSearchPipe
 
         return RoutingSearchPipe(
+            logging_provider=self.providers.logging,
             search_pipes={
                 "vanilla": vanilla_vector_search_pipe,
                 "hyde": hyde_search_pipe,
@@ -459,6 +476,7 @@ class R2RPipeFactory:
         from core.pipes import KGTriplesExtractionPipe
 
         return KGTriplesExtractionPipe(
+            logging_provider=self.providers.logging,
             llm_provider=self.providers.llm,
             database_provider=self.providers.database,
             config=AsyncPipe.PipeConfig(name="kg_triples_extraction_pipe"),
@@ -468,6 +486,7 @@ class R2RPipeFactory:
         from core.pipes import KGStoragePipe
 
         return KGStoragePipe(
+            logging_provider=self.providers.logging,
             database_provider=self.providers.database,
             config=AsyncPipe.PipeConfig(name="kg_storage_pipe"),
         )
@@ -476,6 +495,7 @@ class R2RPipeFactory:
         from core.pipes import KGSearchSearchPipe
 
         return KGSearchSearchPipe(
+            logging_provider=self.providers.logging,
             database_provider=self.providers.database,
             llm_provider=self.providers.llm,
             embedding_provider=self.providers.embedding,
@@ -489,6 +509,7 @@ class R2RPipeFactory:
             from core.pipes import StreamingSearchRAGPipe
 
             return StreamingSearchRAGPipe(
+                logging_provider=self.providers.logging,
                 llm_provider=self.providers.llm,
                 database_provider=self.providers.database,
                 config=GeneratorPipe.PipeConfig(
@@ -499,6 +520,7 @@ class R2RPipeFactory:
             from core.pipes import SearchRAGPipe
 
             return SearchRAGPipe(
+                logging_provider=self.providers.logging,
                 llm_provider=self.providers.llm,
                 database_provider=self.providers.database,
                 config=GeneratorPipe.PipeConfig(
@@ -510,6 +532,7 @@ class R2RPipeFactory:
         from core.pipes import KGEntityDescriptionPipe
 
         return KGEntityDescriptionPipe(
+            logging_provider=self.providers.logging,
             database_provider=self.providers.database,
             llm_provider=self.providers.llm,
             embedding_provider=self.providers.embedding,
@@ -520,6 +543,7 @@ class R2RPipeFactory:
         from core.pipes import KGClusteringPipe
 
         return KGClusteringPipe(
+            logging_provider=self.providers.logging,
             database_provider=self.providers.database,
             llm_provider=self.providers.llm,
             embedding_provider=self.providers.embedding,
@@ -530,6 +554,7 @@ class R2RPipeFactory:
         from core.pipes import KGEntityDeduplicationSummaryPipe
 
         return KGEntityDeduplicationSummaryPipe(
+            logging_provider=self.providers.logging,
             database_provider=self.providers.database,
             llm_provider=self.providers.llm,
             embedding_provider=self.providers.embedding,
@@ -540,6 +565,7 @@ class R2RPipeFactory:
         from core.pipes import KGCommunitySummaryPipe
 
         return KGCommunitySummaryPipe(
+            logging_provider=self.providers.logging,
             database_provider=self.providers.database,
             llm_provider=self.providers.llm,
             embedding_provider=self.providers.embedding,
@@ -550,6 +576,7 @@ class R2RPipeFactory:
         from core.pipes import KGEntityDeduplicationPipe
 
         return KGEntityDeduplicationPipe(
+            logging_provider=self.providers.logging,
             database_provider=self.providers.database,
             llm_provider=self.providers.llm,
             embedding_provider=self.providers.embedding,
@@ -562,6 +589,7 @@ class R2RPipeFactory:
         from core.pipes import KGEntityDeduplicationSummaryPipe
 
         return KGEntityDeduplicationSummaryPipe(
+            logging_provider=self.providers.logging,
             database_provider=self.providers.database,
             llm_provider=self.providers.llm,
             embedding_provider=self.providers.embedding,
@@ -574,6 +602,7 @@ class R2RPipeFactory:
         from core.pipes import KGPromptTuningPipe
 
         return KGPromptTuningPipe(
+            logging_provider=self.providers.logging,
             database_provider=self.providers.database,
             llm_provider=self.providers.llm,
             config=AsyncPipe.PipeConfig(name="kg_prompt_tuning_pipe"),
@@ -581,13 +610,18 @@ class R2RPipeFactory:
 
 
 class R2RPipelineFactory:
-    def __init__(self, config: R2RConfig, pipes: R2RPipes):
+    def __init__(
+        self, config: R2RConfig, providers: R2RProviders, pipes: R2RPipes
+    ):
         self.config = config
+        self.providers = providers
         self.pipes = pipes
 
     def create_search_pipeline(self, *args, **kwargs) -> SearchPipeline:
         """factory method to create an ingestion pipeline."""
-        search_pipeline = SearchPipeline()
+        search_pipeline = SearchPipeline(
+            logging_provider=self.providers.logging
+        )
 
         # Add vector search pipes if embedding provider and vector provider is set
         if (
@@ -614,7 +648,7 @@ class R2RPipelineFactory:
             self.pipes.streaming_rag_pipe if stream else self.pipes.rag_pipe
         )
 
-        rag_pipeline = RAGPipeline()
+        rag_pipeline = RAGPipeline(logging_provider=self.providers.logging)
         rag_pipeline.set_search_pipeline(search_pipeline)
         rag_pipeline.add_pipe(rag_pipe)
         return rag_pipeline
@@ -627,10 +661,6 @@ class R2RPipelineFactory:
         *args,
         **kwargs,
     ) -> R2RPipelines:
-        try:
-            self.configure_logging()
-        except Exception as e:
-            logger.warning(f"Error configuring logging: {e}")
         search_pipeline = search_pipeline or self.create_search_pipeline(
             *args, **kwargs
         )
@@ -651,9 +681,6 @@ class R2RPipelineFactory:
                 **kwargs,
             ),
         )
-
-    def configure_logging(self):
-        R2RLoggingProvider.configure(self.config.logging)
 
 
 class R2RAgentFactory:
