+++ py/core/main/services/base.py
@@ -1,6 +1,7 @@
 from abc import ABC
 
-from core.base import R2RLoggingProvider, RunManager
+from core.base import RunManager
+from core.providers.logging.r2r_logging import SqlitePersistentLoggingProvider
 
 from ..abstractions import R2RAgents, R2RPipelines, R2RPipes, R2RProviders
 from ..config import R2RConfig
@@ -15,7 +16,7 @@ class Service(ABC):
         pipelines: R2RPipelines,
         agents: R2RAgents,
         run_manager: RunManager,
-        logging_connection: R2RLoggingProvider,
+        logging_connection: SqlitePersistentLoggingProvider,
     ):
         self.config = config
         self.providers = providers
