+++ py/core/main/services/kg_service.py
@@ -4,7 +4,7 @@ import time
 from typing import AsyncGenerator, Optional
 from uuid import UUID
 
-from core.base import KGExtractionStatus, R2RLoggingProvider, RunManager
+from core.base import KGExtractionStatus, RunManager
 from core.base.abstractions import (
     GenerationConfig,
     KGCreationSettings,
@@ -13,6 +13,7 @@ from core.base.abstractions import (
     KGEntityDeduplicationType,
     R2RException,
 )
+from core.providers.logging.r2r_logging import SqlitePersistentLoggingProvider
 from core.telemetry.telemetry_decorator import telemetry_event
 
 from ..abstractions import R2RAgents, R2RPipelines, R2RPipes, R2RProviders
@@ -40,7 +41,7 @@ class KgService(Service):
         pipelines: R2RPipelines,
         agents: R2RAgents,
         run_manager: RunManager,
-        logging_connection: R2RLoggingProvider,
+        logging_connection: SqlitePersistentLoggingProvider,
     ):
         super().__init__(
             config,
