+++ py/core/pipelines/graph_enrichment.py
@@ -1,10 +1,10 @@
 import logging
 from typing import Optional
 
-from ..base.logging.r2r_logger import R2RLoggingProvider
 from ..base.logging.run_manager import RunManager
 from ..base.pipeline.base_pipeline import AsyncPipeline
 from ..base.pipes.base_pipe import AsyncPipe
+from ..providers.logging.r2r_logging import SqlitePersistentLoggingProvider
 
 logger = logging.getLogger()
 
@@ -16,10 +16,10 @@ class KGEnrichmentPipeline(AsyncPipeline):
 
     def __init__(
         self,
-        pipe_logger: Optional[R2RLoggingProvider] = None,
+        logging_provider: SqlitePersistentLoggingProvider,
         run_manager: Optional[RunManager] = None,
     ):
-        super().__init__(pipe_logger, run_manager)
+        super().__init__(logging_provider, run_manager)
 
     def add_pipe(
         self,
