+++ py/core/pipelines/rag_pipeline.py
@@ -8,11 +8,11 @@ from ..base.abstractions import (
     VectorSearchSettings,
 )
 from ..base.logging import RunType
-from ..base.logging.r2r_logger import R2RLoggingProvider
 from ..base.logging.run_manager import RunManager, manage_run
 from ..base.pipeline.base_pipeline import AsyncPipeline
 from ..base.pipes.base_pipe import AsyncPipe, AsyncState
 from ..base.utils import to_async_generator
+from ..providers.logging.r2r_logging import SqlitePersistentLoggingProvider
 
 logger = logging.getLogger()
 
@@ -22,10 +22,10 @@ class RAGPipeline(AsyncPipeline):
 
     def __init__(
         self,
-        pipe_logger: Optional[R2RLoggingProvider] = None,
+        logging_provider: SqlitePersistentLoggingProvider,
         run_manager: Optional[RunManager] = None,
     ):
-        super().__init__(pipe_logger, run_manager)
+        super().__init__(logging_provider, run_manager)
         self._search_pipeline: Optional[AsyncPipeline] = None
         self._rag_pipeline: Optional[AsyncPipeline] = None
 
@@ -47,7 +47,9 @@ class RAGPipeline(AsyncPipeline):
         self.state = state or AsyncState()
         # TODO - This feels anti-pattern.
         run_manager = (
-            run_manager or self.run_manager or RunManager(self.pipe_logger)
+            run_manager
+            or self.run_manager
+            or RunManager(self.logging_provider)
         )
         async with manage_run(run_manager, RunType.RETRIEVAL):
             if not self._search_pipeline:
@@ -107,7 +109,9 @@ class RAGPipeline(AsyncPipeline):
                 "Only pipes that are part of the RAG pipeline can be added to the RAG pipeline"
             )
         if not self._rag_pipeline:
-            self._rag_pipeline = AsyncPipeline()
+            self._rag_pipeline = AsyncPipeline(
+                logging_provider=self.logging_provider
+            )
         self._rag_pipeline.add_pipe(
             pipe, add_upstream_outputs, *args, **kwargs
         )
