+++ py/core/pipes/ingestion/vector_storage_pipe.py
@@ -2,15 +2,9 @@ import logging
 from typing import Any, AsyncGenerator, Optional
 from uuid import UUID
 
-from core.base import (
-    AsyncState,
-    DatabaseProvider,
-    PipeType,
-    R2RLoggingProvider,
-    StorageResult,
-    VectorEntry,
-)
+from core.base import AsyncState, DatabaseProvider, StorageResult, VectorEntry
 from core.base.pipes.base_pipe import AsyncPipe
+from core.providers.logging.r2r_logging import SqlitePersistentLoggingProvider
 
 logger = logging.getLogger()
 
@@ -23,9 +17,8 @@ class VectorStoragePipe(AsyncPipe[StorageResult]):
         self,
         database_provider: DatabaseProvider,
         config: AsyncPipe.PipeConfig,
+        logging_provider: SqlitePersistentLoggingProvider,
         storage_batch_size: int = 128,
-        pipe_logger: Optional[R2RLoggingProvider] = None,
-        type: PipeType = PipeType.INGESTOR,
         *args,
         **kwargs,
     ):
@@ -34,8 +27,7 @@ class VectorStoragePipe(AsyncPipe[StorageResult]):
         """
         super().__init__(
             config,
-            type,
-            pipe_logger,
+            logging_provider,
             *args,
             **kwargs,
         )
