+++ py/core/pipes/kg/community_summary.py
@@ -14,10 +14,9 @@ from core.base import (
     DatabaseProvider,
     EmbeddingProvider,
     GenerationConfig,
-    PipeType,
-    R2RLoggingProvider,
 )
 from core.base.abstractions import Entity, Triple
+from core.providers.logging.r2r_logging import SqlitePersistentLoggingProvider
 
 logger = logging.getLogger()
 
@@ -33,8 +32,7 @@ class KGCommunitySummaryPipe(AsyncPipe):
         llm_provider: CompletionProvider,
         embedding_provider: EmbeddingProvider,
         config: AsyncPipe.PipeConfig,
-        pipe_logger: Optional[R2RLoggingProvider] = None,
-        type: PipeType = PipeType.OTHER,
+        logging_provider: SqlitePersistentLoggingProvider,
         *args,
         **kwargs,
     ):
@@ -42,8 +40,7 @@ class KGCommunitySummaryPipe(AsyncPipe):
         Initializes the KG clustering pipe with necessary components and configurations.
         """
         super().__init__(
-            pipe_logger=pipe_logger,
-            type=type,
+            logging_provider=logging_provider,
             config=config
             or AsyncPipe.PipeConfig(name="kg_community_summary_pipe"),
         )
