+++ py/core/pipes/kg/storage.py
@@ -8,11 +8,10 @@ from core.base import (
     DatabaseProvider,
     EmbeddingProvider,
     KGExtraction,
-    PipeType,
     R2RDocumentProcessingError,
-    R2RLoggingProvider,
 )
 from core.base.pipes.base_pipe import AsyncPipe
+from core.providers.logging.r2r_logging import SqlitePersistentLoggingProvider
 
 logger = logging.getLogger()
 
@@ -26,9 +25,8 @@ class KGStoragePipe(AsyncPipe):
         self,
         database_provider: DatabaseProvider,
         config: AsyncPipe.PipeConfig,
+        logging_provider: SqlitePersistentLoggingProvider,
         storage_batch_size: int = 1,
-        pipe_logger: Optional[R2RLoggingProvider] = None,
-        type: PipeType = PipeType.INGESTOR,
         *args,
         **kwargs,
     ):
@@ -41,8 +39,7 @@ class KGStoragePipe(AsyncPipe):
 
         super().__init__(
             config,
-            type,
-            pipe_logger,
+            logging_provider,
             *args,
             **kwargs,
         )
