+++ py/core/providers/database/collection.py
@@ -325,7 +325,7 @@ class PostgresCollectionHandler(CollectionHandler):
                 id=row["document_id"],
                 collection_ids=[collection_id],
                 user_id=row["user_id"],
-                type=DocumentType(row["type"]),
+                document_type=DocumentType(row["type"]),
                 metadata=json.loads(row["metadata"]),
                 title=row["title"],
                 version=row["version"],
