+++ py/core/providers/database/document.py
@@ -398,7 +398,7 @@ class PostgresDocumentHandler(DocumentHandler):
                     id=row["document_id"],
                     collection_ids=row["collection_ids"],
                     user_id=row["user_id"],
-                    type=DocumentType(row["type"]),
+                    document_type=DocumentType(row["type"]),
                     metadata=json.loads(row["metadata"]),
                     title=row["title"],
                     version=row["version"],
