+++ py/core/providers/logging/__init__.py
@@ -1,0 +1,3 @@
+from .r2r_logging import SqlitePersistentLoggingProvider
+
+__all_ = ["SqlitePersistentLoggingProvider"]
