+++ py/shared/abstractions/document.py
@@ -95,7 +95,7 @@ class Document(R2RSerializable):
     id: UUID = Field(default_factory=uuid4)
     collection_ids: list[UUID]
     user_id: UUID
-    type: DocumentType
+    document_type: DocumentType
     metadata: dict
 
     class Config:
@@ -152,7 +152,7 @@ class DocumentInfo(R2RSerializable):
     id: UUID
     collection_ids: list[UUID]
     user_id: UUID
-    type: DocumentType
+    document_type: DocumentType
     metadata: dict
     title: Optional[str] = None
     version: str
@@ -171,7 +171,7 @@ class DocumentInfo(R2RSerializable):
             "document_id": self.id,
             "collection_ids": self.collection_ids,
             "user_id": self.user_id,
-            "type": self.type,
+            "type": self.document_type,
             "metadata": json.dumps(self.metadata),
             "title": self.title or "N/A",
             "version": self.version,
