+++ py/tests/conftest.py
@@ -13,7 +13,7 @@ from core import (
     CompletionConfig,
     DatabaseConfig,
     EmbeddingConfig,
-    LoggingConfig,
+    PersistentLoggingConfig,
     SqlitePersistentLoggingProvider,
     Vector,
     VectorEntry,
@@ -114,7 +114,7 @@ async def postgres_db_provider(
         id=UUID("9fbe403b-c11c-5aae-8ade-ef22980c3ad1"),
         collection_ids=[UUID("122fdf6a-e116-546b-a8f6-e4cb2e2c0a09")],
         user_id=UUID("00000000-0000-0000-0000-000000000003"),
-        type=DocumentType.PDF,
+        document_type=DocumentType.PDF,
         metadata={},
         title="Test Document for KG",
         version="1.0",
@@ -214,7 +214,7 @@ async def local_logging_provider(app_config):
     unique_id = str(uuid.uuid4())
     logging_path = f"test_{unique_id}.sqlite"
     provider = SqlitePersistentLoggingProvider(
-        LoggingConfig(logging_path=logging_path, app=app_config)
+        PersistentLoggingConfig(logging_path=logging_path, app=app_config)
     )
     await provider._init()
     yield provider
