+++ py/tests/core/providers/logging/test_logging_provider.py
@@ -6,7 +6,7 @@ from uuid import UUID
 import pytest
 
 from core import (
-    LoggingConfig,
+    PersistentLoggingConfig,
     SqlitePersistentLoggingProvider,
     generate_run_id,
 )
